started work on getting the tests to compile/run on windows

+++ source/extensions/filters/network/redis_proxy/conn_pool_impl.cc
@@ -261,7 +261,7 @@ InstanceImpl::ThreadLocalPool::makeRequestToHost(const std::string& host_address
     host_address_map_key = host_address;
   } else {
     const auto ip_port = absl::string_view(host_address).substr(colon_pos + 1);
-    uint64_t ip_port_number;
+    uint32_t ip_port_number;
     if (!absl::SimpleAtoi(ip_port, &ip_port_number) || (ip_port_number > 65535)) {
       return nullptr;
     }
